+++ libraries/kotlinx-metadata/src/kotlinx/metadata/impl/readUtils.kt
@@ -23,13 +23,14 @@ fun ProtoBuf.Annotation.readAnnotation(strings: NameResolver): KmAnnotation =
         }.toMap()
     )
 
-fun ProtoBuf.Annotation.Argument.Value.readAnnotationArgument(strings: NameResolver): KmAnnotationArgument<*>? {
+@OptIn(ExperimentalUnsignedTypes::class)
+fun ProtoBuf.Annotation.Argument.Value.readAnnotationArgument(strings: NameResolver): KmAnnotationArgument? {
     if (Flags.IS_UNSIGNED[flags]) {
         return when (type) {
-            BYTE -> KmAnnotationArgument.UByteValue(intValue.toByte())
-            SHORT -> KmAnnotationArgument.UShortValue(intValue.toShort())
-            INT -> KmAnnotationArgument.UIntValue(intValue.toInt())
-            LONG -> KmAnnotationArgument.ULongValue(intValue)
+            BYTE -> KmAnnotationArgument.UByteValue(intValue.toByte().toUByte())
+            SHORT -> KmAnnotationArgument.UShortValue(intValue.toShort().toUShort())
+            INT -> KmAnnotationArgument.UIntValue(intValue.toInt().toUInt())
+            LONG -> KmAnnotationArgument.ULongValue(intValue.toULong())
             else -> error("Cannot read value of unsigned type: $type")
         }
     }