+++ native/commonizer/src/org/jetbrains/kotlin/commonizer/metadata/CirDeserializers.kt
@@ -11,10 +11,10 @@ import kotlinx.metadata.klib.annotations
 import kotlinx.metadata.klib.compileTimeValue
 import kotlinx.metadata.klib.getterAnnotations
 import kotlinx.metadata.klib.setterAnnotations
-import org.jetbrains.kotlin.descriptors.*
 import org.jetbrains.kotlin.commonizer.cir.*
 import org.jetbrains.kotlin.commonizer.mergedtree.CirProvided
 import org.jetbrains.kotlin.commonizer.utils.*
+import org.jetbrains.kotlin.descriptors.*
 import org.jetbrains.kotlin.types.Variance
 
 object CirDeserializers {
@@ -45,7 +45,7 @@ object CirDeserializers {
             isMarkedNullable = false
         )
 
-        val allValueArguments: Map<String, KmAnnotationArgument<*>> = source.arguments
+        val allValueArguments: Map<String, KmAnnotationArgument> = source.arguments
         if (allValueArguments.isEmpty())
             return CirAnnotation.createInterned(type = type, constantValueArguments = emptyMap(), annotationValueArguments = emptyMap())
 
@@ -55,7 +55,7 @@ object CirDeserializers {
         allValueArguments.forEach { (name, constantValue) ->
             val cirName = CirName.create(name)
             if (constantValue is KmAnnotationArgument.AnnotationValue)
-                annotationValueArguments[cirName] = annotation(source = constantValue.value, typeResolver)
+                annotationValueArguments[cirName] = annotation(source = constantValue.annotation, typeResolver)
             else
                 constantValueArguments[cirName] = constantValue(
                     constantValue = constantValue,
@@ -205,7 +205,7 @@ object CirDeserializers {
         )
 
     private fun constantValue(
-        constantValue: KmAnnotationArgument<*>?,
+        constantValue: KmAnnotationArgument?,
         constantName: CirName? = null,
         owner: Any,
     ): CirConstantValue<*> = constantValue(
@@ -213,8 +213,9 @@ object CirDeserializers {
         location = { "${owner::class.java}, $owner" + constantName?.toString()?.let { "[$it]" } }
     )
 
+    @OptIn(ExperimentalUnsignedTypes::class)
     private fun constantValue(
-        constantValue: KmAnnotationArgument<*>?,
+        constantValue: KmAnnotationArgument?,
         location: () -> String
     ): CirConstantValue<*> = when (constantValue) {
         null -> CirConstantValue.NullValue
@@ -242,7 +243,7 @@ object CirDeserializers {
         )
 
         is KmAnnotationArgument.ArrayValue -> CirConstantValue.ArrayValue(
-            constantValue.value.compactMapIndexed { index, innerConstantValue ->
+            constantValue.elements.compactMapIndexed { index, innerConstantValue ->
                 constantValue(
                     constantValue = innerConstantValue,
                     location = { "${location()}[$index]" }