+++ native/commonizer/src/org/jetbrains/kotlin/commonizer/metadata/CirSerializers.kt
@@ -8,13 +8,13 @@ package org.jetbrains.kotlin.commonizer.metadata
 import kotlinx.metadata.*
 import kotlinx.metadata.klib.*
 import org.jetbrains.kotlin.backend.common.serialization.metadata.DynamicTypeDeserializer
-import org.jetbrains.kotlin.descriptors.CallableMemberDescriptor
 import org.jetbrains.kotlin.commonizer.cir.*
 import org.jetbrains.kotlin.commonizer.mergedtree.*
 import org.jetbrains.kotlin.commonizer.metadata.TypeAliasExpansion.*
 import org.jetbrains.kotlin.commonizer.utils.DEFAULT_SETTER_VALUE_NAME
 import org.jetbrains.kotlin.commonizer.utils.SPECIAL_CLASS_WITHOUT_SUPERTYPES_CLASS_NAMES
 import org.jetbrains.kotlin.commonizer.utils.compactMap
+import org.jetbrains.kotlin.descriptors.CallableMemberDescriptor
 import org.jetbrains.kotlin.types.Variance
 
 internal fun CirModule.serializeModule(
@@ -195,7 +195,7 @@ internal fun CirFunction.serializeFunction(
 }
 
 private fun CirAnnotation.serializeAnnotation(): KmAnnotation {
-    val arguments = LinkedHashMap<String, KmAnnotationArgument<*>>(constantValueArguments.size + annotationValueArguments.size, 1F)
+    val arguments = LinkedHashMap<String, KmAnnotationArgument>(constantValueArguments.size + annotationValueArguments.size, 1F)
 
     constantValueArguments.forEach { (name: CirName, value: CirConstantValue<*>) ->
         arguments[name.name] = value.serializeConstantValue()
@@ -212,7 +212,8 @@ private fun CirAnnotation.serializeAnnotation(): KmAnnotation {
     )
 }
 
-private fun CirConstantValue<*>.serializeConstantValue(): KmAnnotationArgument<*>? = when (this) {
+@OptIn(ExperimentalUnsignedTypes::class)
+private fun CirConstantValue<*>.serializeConstantValue(): KmAnnotationArgument? = when (this) {
     is CirConstantValue.StringValue -> KmAnnotationArgument.StringValue(value)
     is CirConstantValue.CharValue -> KmAnnotationArgument.CharValue(value)